requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
numpy
tensorflow-cpu  # or 'torch' depending on your model
av